1318 - move file-upload folder to common
